chatbot/requirements.txt

flask
openai
sympy
python-dotenv
